requirements.txt
# Core
streamlit>=1.32.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Database drivers
psycopg2-binary>=2.9.9  # PostgreSQL
mysql-connector-python>=8.2.0  # MySQL

# Statistics
scipy>=1.12.0
statsmodels>=0.14.0

# Visualization
plotly>=5.19.0
altair>=5.2.0

# Utilities
pydantic>=2.6.0
python-dotenv>=1.0.0
